REFACTOR: Proptypes 추가 / ModalSearchInput atom 폴더 위치 변경
