+++ src/atom/useSearchInput.js
@@ -1,4 +1,7 @@
-/* eslint-disable import/prefer-default-export */
 import { atom } from "recoil";
 
 export const useSearchInput = atom({ key: "useSearchInput", default: "" });
+export const useModalSearchInput = atom({
+  key: "useModalSearchInput",
+  default: "",
+});
