+++ src/component/reservedloan/ReservedLoan.js
@@ -7,7 +7,7 @@ import "../../css/ReservedLoan.css";
 import SubTitle from "../utils/SubTitle";
 import ModalPagination from "../rent/ModalPagination";
 import InquireBoxTitle from "../utils/InquireBoxTitle";
-import { useModalSearchInput } from "../utils/ModalSearchBar";
+import { useModalSearchInput } from "../../atom/useSearchInput";
 import Login from "../../img/login_icon.svg";
 import ReservedFilter from "./ReservedFilter";
 import ReservedTableList from "./ReservedTableList";
